+++ crates/console/src/components/mod.rs
@@ -1,9 +1,11 @@
-mod data;
 mod entry;
 mod layout;
+mod list;
+mod map;
 mod styled_text;
 
-pub use data::*;
 pub use entry::*;
 pub use layout::*;
+pub use list::*;
+pub use map::*;
 pub use styled_text::*;
